fixed all initial compiling errors that arose from reducer and action refactoring
