+++ src/reducers/farmerReducer.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { FARMER_SEARCH_START, FARMER_SEARCH_SUCCESS, FARMER_SEARCH_FAILURE,
         ADD_FARMER_START, ADD_FARMER_SUCCESS, ADD_FARMER_FAILURE,
         DELETE_FARMER_START, DELETE_FARMER_SUCCESS, DELETE_FARMER_FAILURE 
-} from "../actions/FarmerSearch";
+} from "../actions";
 
 const initialState = {
     deleteStart: false,
