+++ src/reducers/organizationReducer.js
@@ -1,7 +1,7 @@
 import { 
     GET_ORGANIZATION, GET_ORGANIZATION_SUCCESS, GET_ORGANIZATION_FAILURE,
     ADD_ORGANIZATION_START, ADD_ORGANIZATION_SUCCESS, ADD_ORGANIZATION_FAILURE,
-    ORGANIZATION_SEARCH_START, ORGANIZATION_SEARCH_SUCCESS, ORGANIZATION_SEARCH_FAILURE, ADD_ORGANIZATION_START
+    ORGANIZATION_SEARCH_START, ORGANIZATION_SEARCH_SUCCESS, ORGANIZATION_SEARCH_FAILURE
 } from '../actions'
 
 const initialState = {
